feat: user registration and password encryption

+++ services/node/src/application/src/authorization/user_token_payload.rs
@@ -2,5 +2,5 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct UserTokenPayload {
-    pub user_id: usize,
+    pub user_id: u64,
 }
